refactor: ThemeToggleSwitch 컴포넌트를 common 폴더로 이동 #19
